+++ examples/rest-client-example/src/main/java/com/kevshah/example/restclientexample/Application.java
@@ -1,0 +1,13 @@
+package com.kevshah.example.restclientexample;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
+